Frontend client has its own container now.
The index.html is served by the fastify-static plugin.
We have to adapt the frontend to make calls directly to the gateway service now.
Had to make some changes to building of the frontend. Since now every containers has just a part of the application, there is no need for the /frontend after build directory